+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Version 8.0.0-dev
 - BC break: removed support for old way of providing 1 parameter to a test method via data provided
 - BC break: removed support for using conditions with attribute Skip
 - added event ExtensionsLoaded
-- BC break: added methods getName amd onExtensionsLoaded to interface ITesterExtension
+- BC break: added methods getName and getSubscribedEvents to interface ITesterExtension, removed methods on* (it is now an event subscriber)
 - InfoExtension (when added to automated tests runner) prints active extensions
 
 Version 7.3.1